+++ RobotArdu/resources/libraries/RobertaFunctions/RobertaFunctions.h
@@ -17,8 +17,8 @@
 class RobertaFunctions
 {
   public:   
-		void moveTime(int speedL,int speedR, long time);
-		void moveTimePID(int speedL,int speedR, long time);
+		void moveTime(int speedL,int speedR, double time);
+		void moveTimePID(int speedL,int speedR, double time);
 		void lcdClear();
 		int ultrasonicDistance(int port);
 		byte *colorSensorRGB(byte colors[],int port);
